requirements.txt
python3-linkedin==1.0.2